move includes into UserCompactTrasformer; apply permission to is_admin
