+++ app/Transformers/UserTransformer.php
@@ -5,15 +5,14 @@
 
 namespace App\Transformers;
 
-use App\Models\Beatmap;
 use App\Models\User;
-use League\Fractal;
 
 class UserTransformer extends UserCompactTransformer
 {
     protected $defaultIncludes = [
         'country',
         'cover',
+        'is_admin',
         'is_bng',
         'is_full_bn',
         'is_gmt',
@@ -25,6 +24,7 @@ class UserTransformer extends UserCompactTransformer
 
     protected $permissions = [
         'friends' => 'IsNotOAuth',
+        'is_admin' => 'IsNotOAuth',
         'is_bng' => 'IsNotOAuth',
         'is_full_bn' => 'IsNotOAuth',
         'is_gmt' => 'IsNotOAuth',
@@ -33,41 +33,6 @@ class UserTransformer extends UserCompactTransformer
         'is_nat' => 'IsNotOAuth',
         'is_restricted' => 'IsNotOAuth',
     ];
-
-    public function __construct()
-    {
-        static $includes;
-
-        if (!isset($includes)) {
-            $includes = array_merge($this->availableIncludes, [
-                'account_history',
-                'active_tournament_banner',
-                'badges',
-                'blocks',
-                'defaultStatistics',
-                'favourite_beatmapset_count',
-                'follower_count',
-                'friends',
-                'graveyard_beatmapset_count',
-                'is_admin',
-                'loved_beatmapset_count',
-                'monthly_playcounts',
-                'page',
-                'previous_usernames',
-                'ranked_and_approved_beatmapset_count',
-                'rankHistory', // TODO: should be changed to rank_history
-                'replays_watched_counts',
-                'scores_first_count',
-                'statistics',
-                'unranked_beatmapset_count',
-                'unread_pm_count',
-                'user_achievements',
-                'user_preferences',
-            ]);
-        }
-
-        $this->availableIncludes = $includes;
-    }
 
     public function transform(User $user)
     {
@@ -105,78 +70,9 @@ class UserTransformer extends UserCompactTransformer
         ]);
     }
 
-    public function includeAccountHistory(User $user)
-    {
-        $histories = $user->accountHistories()->recent();
-
-        if (!priv_check('UserSilenceShowExtendedInfo')->can()) {
-            $histories->default();
-        } else {
-            $histories->with('actor');
-        }
-
-        return $this->collection(
-            $histories->get(),
-            new UserAccountHistoryTransformer()
-        );
-    }
-
-    public function includeActiveTournamentBanner(User $user)
-    {
-        return $this->item($user->profileBanners()->active(), new ProfileBannerTransformer);
-    }
-
-    public function includeBadges(User $user)
-    {
-        return $this->collection(
-            $user->badges()->orderBy('awarded', 'DESC')->get(),
-            new UserBadgeTransformer
-        );
-    }
-
-    public function includeDefaultStatistics(User $user)
-    {
-        $stats = $user->statistics($user->playmode);
-
-        return $this->item($stats, new UserStatisticsTransformer);
-    }
-
-    public function includeFavouriteBeatmapsetCount(User $user)
-    {
-        return $this->primitive($user->profileBeatmapsetsFavourite()->count());
-    }
-
-    public function includeBlocks(User $user)
-    {
-        return $this->collection(
-            $user->relations()->blocks()->get(),
-            new UserRelationTransformer()
-        );
-    }
-
-    public function includeFollowerCount(User $user)
-    {
-        return $this->primitive($user->followerCount());
-    }
-
-    public function includeFriends(User $user)
-    {
-        return $this->collection(
-            $user->relations()->friends()->withMutual()->get(),
-            new UserRelationTransformer()
-        );
-    }
-
-    public function includeGraveyardBeatmapsetCount(User $user)
-    {
-        return $this->primitive($user->profileBeatmapsetsGraveyard()->count());
-    }
-
     public function includeIsAdmin(User $user)
     {
-        return $this->primitive($user->isAdmin(), function ($flag) {
-            return $flag;
-        });
+        return $this->primitive($user->isAdmin());
     }
 
     public function includeIsBng(User $user)
@@ -212,107 +108,5 @@ class UserTransformer extends UserCompactTransformer
     public function includeIsRestricted(User $user)
     {
         return $this->primitive($user->isRestricted());
-    }
-
-    public function includeLovedBeatmapsetCount(User $user)
-    {
-        return $this->primitive($user->profileBeatmapsetsLoved()->count());
-    }
-
-    public function includeMonthlyPlaycounts(User $user)
-    {
-        return $this->collection(
-            $user->monthlyPlaycounts,
-            new UserMonthlyPlaycountTransformer
-        );
-    }
-
-    public function includePage(User $user)
-    {
-        return $this->item($user, function ($user) {
-            if ($user->userPage !== null) {
-                return [
-                    'html' => $user->userPage->bodyHTML(['withoutImageDimensions' => true, 'modifiers' => ['profile-page']]),
-                    'raw' => $user->userPage->bodyRaw,
-                ];
-            } else {
-                return ['html' => '', 'raw' => ''];
-            }
-        });
-    }
-
-    public function includePreviousUsernames(User $user)
-    {
-        return $this->primitive($user->previousUsernames()->unique()->values()->toArray());
-    }
-
-    public function includeRankedAndApprovedBeatmapsetCount(User $user)
-    {
-        return $this->primitive($user->profileBeatmapsetsRankedAndApproved()->count());
-    }
-
-    public function includeRankHistory(User $user, Fractal\ParamBag $params)
-    {
-        $mode = $params->get('mode')[0];
-
-        $rankHistoryData = $user->rankHistories()
-            ->where('mode', Beatmap::modeInt($mode))
-            ->first();
-
-        return $rankHistoryData === null
-            ? $this->primitive(null)
-            : $this->item($rankHistoryData, new RankHistoryTransformer);
-    }
-
-    public function includeReplaysWatchedCounts(User $user)
-    {
-        return $this->collection(
-            $user->replaysWatchedCounts,
-            new UserReplaysWatchedCountTransformer
-        );
-    }
-
-    public function includeScoresFirstCount(User $user, Fractal\ParamBag $params)
-    {
-        $mode = $params->get('mode')[0];
-
-        return $this->primitive($user->scoresFirst($mode, true)->visibleUsers()->count());
-    }
-
-    public function includeStatistics(User $user, Fractal\ParamBag $params)
-    {
-        $stats = $user->statistics($params->get('mode')[0]);
-
-        return $this->item($stats, new UserStatisticsTransformer);
-    }
-
-    public function includeUnrankedBeatmapsetCount(User $user)
-    {
-        return $this->primitive($user->profileBeatmapsetsUnranked()->count());
-    }
-
-    public function includeUnreadPmCount(User $user)
-    {
-        return $this->primitive($user->notificationCount());
-    }
-
-    public function includeUserAchievements(User $user)
-    {
-        return $this->collection(
-            $user->userAchievements()->orderBy('date', 'desc')->get(),
-            new UserAchievementTransformer()
-        );
-    }
-
-    public function includeUserPreferences(User $user)
-    {
-        $customization = $user->userProfileCustomization ?? $user->userProfileCustomization()->make();
-
-        return $this->primitive([
-            'ranking_expanded' => $customization->ranking_expanded,
-            'user_list_filter' => $customization->user_list_filter,
-            'user_list_sort' => $customization->user_list_sort,
-            'user_list_view' => $customization->user_list_view,
-        ]);
     }
 }
